Muevo los precios a una variable externa
