+++ src/container/Offices/Offices.jsx
@@ -3,71 +3,12 @@ import { Link } from 'react-scroll';
 
 import 'react-multi-carousel/lib/styles.css';
 import { PeopleIcon } from '../../components/Icons';
+import PRICES from '../../data/prices';
 
-const PLANS = [
-  {
-    id: 4,
-    title: 'Escritorio Dedicado',
-    imageURL: 'images/offices/fixed-spot.jpg',
-    tag: 'office',
-    tagIcon: 'location',
-    capacity: '2 personas',
-    billingPeriod: 'Mes',
-    price: 29000,
-    description:
-      'Espacio de trabajo exclusivo mensual para 2 personas, con escritorios dedicados y mobiliario ergonómico.',
-    features: [
-      'Espacio para 2 personas',
-      'Uso exclusivo y fijo',
-      'Mobiliario ergonómico',
-      'Conexión Wi-Fi de alta velocidad',
-      'Acceso a áreas comunes',
-    ],
-  },
-  {
-    id: 5,
-    title: 'Oficinas',
-    imageURL: 'images/offices/office-6p.jpg',
-    tag: 'office',
-    tagIcon: 'people',
-    capacity: 'de 4 a 6 personas',
-    billingPeriod: 'Mes',
-    priceLabel: 'desde',
-    price: 80000,
-    description:
-      'Oficinas privadas para 4 o 6 personas, con ambiente cómodo, equipamiento ergonómico y espacio colaborativo.',
-  },
-  {
-    id: 6,
-    title: 'Salas de reuniones',
-    imageURL: 'images/offices/meetings.jpg',
-    tag: 'office',
-    capacity: 'de 4 a 16 personas',
-    tagIcon: 'people',
-    billingPeriod: 'Hora',
-    priceLabel: 'desde',
-    price: 2500,
-    description:
-      'Organiza reuniones en espacios privados para 4, 6 o 16 personas, con comodidad y equipamiento adecuado.',
-  },
-  {
-    id: 7,
-    title: 'Oficina Virtual',
-    tag: '',
-    tagIcon: 'phone',
-    billingPeriod: 'Mes',
-    price: 5800,
-    description:
-      'Proporcionamos domicilio comercial, gestión de correspondencia y atención a tus clientes de manera telefónica o virtual via e-mail.',
-  },
-];
-
-const ImageCard = ({
-  title,
+const OfficeCard = ({
+  name,
   imageURL,
-  tag,
   capacity,
-  tagIcon,
   billingPeriod,
   price,
   description,
@@ -81,14 +22,13 @@ const ImageCard = ({
         alt='avatar'
       />
 
-      <div className='flex items-center px-6 py-3 bg-gray-900'>
+      <div className='flex items-center px-6 py-3 bg-black'>
         <PeopleIcon />
-
         <h1 className='mx-3 text-lg font-semibold text-white'>{capacity}</h1>
       </div>
 
       <div className='px-6 py-4 flex flex-col justify-end'>
-        <h1 className='text-xl font-semibold text-gray-800'>{title}</h1>
+        <h1 className='text-xl font-semibold text-gray-800'>{name}</h1>
 
         <p className='py-2 text-gray-700 mb-8 flex-grow'>{description}</p>
         <div className='flex flex-col min-h-16 justify-end'>
@@ -115,7 +55,7 @@ const ImageCard = ({
 };
 
 const Offices = () => {
-  const officesPlans = PLANS.filter(({ tag }) => tag === 'office');
+  const offices = PRICES.filter(({ type }) => type === 'office');
   return (
     <section id='oficinas' className='bg-gray-100 pb-14'>
       <div className='container px-6 py-8 mx-auto'>
@@ -133,8 +73,8 @@ const Offices = () => {
         </div>
 
         <div className='grid gap-6 mt-16 -mx-6 sm:gap-8 sm:grid-cols-2 lg:grid-cols-3 px-6'>
-          {officesPlans.map((plan, index) => (
-            <ImageCard {...plan} key={index} />
+          {offices.map((office, index) => (
+            <OfficeCard {...office} key={index} />
           ))}
         </div>
       </div>
