+++ src/container/Plans/PlansV2.jsx
@@ -2,132 +2,14 @@ import React from 'react';
 import { Link } from 'react-scroll';
 
 import 'react-multi-carousel/lib/styles.css';
-import {
-  ClockIcon,
-  LocationIcon,
-  PeopleIcon,
-  PhoneIcon,
-} from '../../components/Icons';
 
-const PLANS = [
-  {
-    id: 1,
-    title: 'Pase por el día',
-    tag: 'open-space',
-    tagIcon: 'clock',
-    billingPeriod: 'Día',
-    price: 3000,
-    description:
-      'Disfrutá de una jornada laboral completa, aprovechando todos los beneficios que Workbox brinda.',
-    features: [
-      'Todas las áreas comunes incluidas',
-      'Conexión Wi-Fi de alta velocidad',
-      'Ambiente de trabajo colaborativo',
-      'Cocina con café y snacks',
-    ],
-  },
-  {
-    id: 3,
-    title: 'Pase mensual',
-    tag: 'open-space',
-    tagIcon: 'clock',
-    billingPeriod: 'Mes',
-    price: 21000,
-    description:
-      'Acceso ilimitado en horario de atención para coworkers que desean aprovechar al máximo nuestras instalaciones.',
-    features: [
-      'Todas las áreas comunes incluidas',
-      'Conexión Wi-Fi de alta velocidad',
-      'Ambiente de trabajo colaborativo',
-      'Cocina con café y snacks',
-      '1 hora de sala de reuniones al mes',
-    ],
-  },
-  {
-    id: 4,
-    title: 'Escritorio Dedicado',
-    imageURL: 'images/offices/fixed-spot.jpg',
-    tag: 'office',
-    tagIcon: 'location',
-    capacity: '2 personas',
-    billingPeriod: 'Mes',
-    price: 29000,
-    description:
-      'Espacio semi-privado fijo y de uso exclusivo para 2 personas, equipado con el mejor mobiliario.',
-    features: [
-      'Espacio para 2 personas',
-      'Uso exclusivo y fijo',
-      'Mobiliario ergonómico',
-      'Conexión Wi-Fi de alta velocidad',
-      'Acceso a áreas comunes',
-    ],
-  },
-  {
-    id: 5,
-    title: 'Oficinas',
-    imageURL: 'images/offices/office-6p.jpg',
-    tag: 'office',
-    tagIcon: 'people',
-    capacity: 'de 4 a 6 personas',
-    billingPeriod: 'Mes',
-    priceLabel: 'desde',
-    price: 80000,
-    description:
-      'Disfrutá de un ambiente privado y agradable para 6 personas, con equipamiento ergonómico premium.',
-  },
-  {
-    id: 6,
-    title: 'Salas de reuniones',
-    imageURL: 'images/offices/meetings.jpg',
-    tag: 'office',
-    capacity: 'de 4 a 16 personas',
-    tagIcon: 'people',
-    billingPeriod: 'Hora',
-    priceLabel: 'desde',
-    price: 2500,
-    description:
-      'Programa reuniones privadas con tu equipo o clientes en un espacio cómodo y adecuadamente equipado.',
-  },
-  {
-    id: 7,
-    title: 'Oficina Virtual',
-    tag: '',
-    tagIcon: 'phone',
-    billingPeriod: 'Mes',
-    price: 5800,
-    description:
-      'Proporcionamos domicilio comercial, gestión de correspondencia y atención a tus clientes de manera telefónica o virtual via e-mail.',
-  },
-];
+import PRICES from '../../data/prices';
 
-const FeatureCard = ({
-  title,
-  tag,
-  tagIcon,
-  price,
-  description,
-  billingPeriod,
-  features,
-}) => {
-  const getTagIcon = ({ tagIcon, size }) => {
-    switch (tagIcon) {
-      default:
-        return null;
-      case 'clock':
-        return <ClockIcon size={size} />;
-      case 'people':
-        return <PeopleIcon />;
-      case 'phone':
-        return <PhoneIcon />;
-      case 'location':
-        return <LocationIcon size={size} />;
-    }
-  };
-
+const FeatureCard = ({ name, price, description, billingPeriod, features }) => {
   return (
     <div className='max-w-md border rounded-lg'>
       <div className='p-6'>
-        <h1 className='text-2xl font-medium text-gray-700'>{title}</h1>
+        <h1 className='text-2xl font-bold text-gray-700'>{name}</h1>
 
         <h4 className='text-4xl font-semibold text-gray-800 opacity-90 mt-6'>
           <span className='text-oldGold'>$</span>
@@ -183,7 +65,7 @@ const FeatureCard = ({
 };
 
 const PlansV2 = () => {
-  const openSpacePlans = PLANS.filter(({ tag }) => tag === 'open-space');
+  const openSpacePlans = PRICES.filter(({ type }) => type === 'open-space');
   return (
     <section id='planes' className='bg-white pb-14'>
       <div className='container px-6 py-8 mx-auto'>
