+++ src/container/index.js
@@ -1,8 +1,8 @@
-import Main from "./Main/Main";
-import Gallery from "./Gallery/Gallery";
-import Footer from "./Footer/Footer";
-import Services from "./Services/Services";
-import Contact from "./Contact/Contact";
-import Plans from "./Plans/Plans";
-
-export { Main, Gallery, Footer, Services, Contact, Plans };
+export { default as Contact } from './Contact/Contact';
+export { default as Footer } from './Footer/Footer';
+export { default as Gallery } from './Gallery/Gallery';
+export { default as Main } from './Main/Main';
+export { default as Offices } from './Offices/Offices';
+export { default as PlansV2 } from './Plans/PlansV2';
+export { default as Services } from './Services/Services';
+export { default as VirtualOffice } from './VirtualOffice/VirtualOffice';
