+++ src/data/prices.js
@@ -1,0 +1,177 @@
+const PLANS = [
+  {
+    id: 1,
+    title: 'Pase por el día',
+    tag: 'open-space',
+    tagIcon: 'clock',
+    billingPeriod: 'Día',
+    price: 3000,
+    description:
+      'Disfrutá de una jornada laboral completa, aprovechando todos los beneficios que Workbox brinda.',
+    features: [
+      'Todas las áreas comunes incluidas',
+      'Conexión Wi-Fi de alta velocidad',
+      'Ambiente de trabajo colaborativo',
+      'Cocina con café y snacks',
+    ],
+  },
+  {
+    id: 3,
+    title: 'Pase mensual',
+    tag: 'open-space',
+    tagIcon: 'clock',
+    billingPeriod: 'Mes',
+    price: 21000,
+    description:
+      'Acceso ilimitado en horario de atención para coworkers que desean aprovechar al máximo nuestras instalaciones.',
+    features: [
+      'Todas las áreas comunes incluidas',
+      'Conexión Wi-Fi de alta velocidad',
+      'Ambiente de trabajo colaborativo',
+      'Cocina con café y snacks',
+      '1 hora de sala de reuniones al mes',
+    ],
+  },
+  {
+    id: 4,
+    title: 'Escritorio Dedicado',
+    imageURL: 'images/offices/fixed-spot.jpg',
+    tag: 'office',
+    tagIcon: 'location',
+    capacity: '2 personas',
+    billingPeriod: 'Mes',
+    price: 29000,
+    description:
+      'Espacio semi-privado fijo y de uso exclusivo para 2 personas, equipado con el mejor mobiliario.',
+    features: [
+      'Espacio para 2 personas',
+      'Uso exclusivo y fijo',
+      'Mobiliario ergonómico',
+      'Conexión Wi-Fi de alta velocidad',
+      'Acceso a áreas comunes',
+    ],
+  },
+  {
+    id: 5,
+    title: 'Oficinas',
+    imageURL: 'images/offices/office-6p.jpg',
+    tag: 'office',
+    tagIcon: 'people',
+    capacity: 'de 4 a 6 personas',
+    billingPeriod: 'Mes',
+    priceLabel: 'desde',
+    price: 80000,
+    description:
+      'Disfrutá de un ambiente privado y agradable para 6 personas, con equipamiento ergonómico premium.',
+  },
+  {
+    id: 6,
+    title: 'Salas de reuniones',
+    imageURL: 'images/offices/meetings.jpg',
+    tag: 'office',
+    capacity: 'de 4 a 16 personas',
+    tagIcon: 'people',
+    billingPeriod: 'Hora',
+    priceLabel: 'desde',
+    price: 2500,
+    description:
+      'Programa reuniones privadas con tu equipo o clientes en un espacio cómodo y adecuadamente equipado.',
+  },
+  {
+    id: 7,
+    title: 'Oficina Virtual',
+    tag: '',
+    tagIcon: 'phone',
+    billingPeriod: 'Mes',
+    price: 5800,
+    description:
+      'Proporcionamos domicilio comercial, gestión de correspondencia y atención a tus clientes de manera telefónica o virtual via e-mail.',
+  },
+];
+
+const PRICES = [
+  {
+    id: 1,
+    name: 'Pase por el día',
+    type: 'open-space',
+    billingPeriod: 'Día',
+    price: 3000,
+    description:
+      'Disfrutá de una jornada laboral completa, aprovechando todos los beneficios que Workbox brinda.',
+    features: [
+      'Todas las áreas comunes incluidas',
+      'Conexión Wi-Fi de alta velocidad',
+      'Ambiente de trabajo colaborativo',
+      'Cocina con café y snacks',
+    ],
+  },
+  {
+    id: 2,
+    name: 'Pase mensual',
+    type: 'open-space',
+    billingPeriod: 'Mes',
+    price: 21000,
+    description:
+      'Acceso ilimitado en horario de atención para coworkers que desean aprovechar al máximo nuestras instalaciones.',
+    features: [
+      'Todas las áreas comunes incluidas',
+      'Conexión Wi-Fi de alta velocidad',
+      'Ambiente de trabajo colaborativo',
+      'Cocina con café y snacks',
+      '1 hora de sala de reuniones al mes',
+    ],
+  },
+  {
+    id: 3,
+    name: 'Escritorio Dedicado',
+    imageURL: 'images/offices/fixed-spot.jpg',
+    type: 'office',
+    capacity: '2 personas',
+    billingPeriod: 'Mes',
+    price: 29000,
+    description:
+      'Espacio de trabajo exclusivo mensual para 2 personas, con escritorios dedicados y mobiliario ergonómico.',
+    features: [
+      'Espacio para 2 personas',
+      'Uso exclusivo y fijo',
+      'Mobiliario ergonómico',
+      'Conexión Wi-Fi de alta velocidad',
+      'Acceso a áreas comunes',
+    ],
+  },
+  {
+    id: 4,
+    name: 'Oficinas',
+    imageURL: 'images/offices/office-6p.jpg',
+    type: 'office',
+    capacity: 'de 4 a 6 personas',
+    billingPeriod: 'Mes',
+    priceLabel: 'desde',
+    price: 80000,
+    description:
+      'Oficinas privadas para 4 o 6 personas, con ambiente cómodo, equipamiento ergonómico y espacio colaborativo.',
+  },
+  {
+    id: 5,
+    name: 'Salas de reuniones',
+    imageURL: 'images/offices/meetings.jpg',
+    type: 'office',
+    capacity: 'de 4 a 16 personas',
+    billingPeriod: 'Hora',
+    priceLabel: 'desde',
+    price: 2500,
+    description:
+      'Organiza reuniones en espacios privados para 4, 6 o 16 personas, con comodidad y equipamiento adecuado.',
+  },
+  {
+    id: 6,
+    name: 'Oficina Virtual',
+    type: 'virtual-office',
+    description:
+      'Brindamos domicilio comercial, gestión de correspondencia y atención telefónica o virtual por e-mail para tus clientes.',
+    price: 2900,
+    billingPeriod: 'Mes',
+  },
+];
+
+export default PRICES;
